+++ mspayment/src/main/java/br/com/bancoamericano/mspayment/model/dto/CalculateRuleResponse.java
@@ -9,5 +9,5 @@ import java.math.BigDecimal;
 @Data
 public class CalculateRuleResponse {
 
-    private BigDecimal total;
+    private Long total;
 }
